requirements.txt
tensorflow
keras
numpy
matplotlib
Pillow
streamlit
scipy
